{"schema":"decodex/commit/1","summary":"Split MCP types","authority":"XY-1115"}
